requirements.txt
# These are dependencies specified for a Python project
setuptools~=68.2.2  # Dependency for package setup
pillow~=10.2.0  # Imaging library
requests~=2.31.0  # HTTP library for making requests
instagrapi~=2.0.3  # Instagram API client
asyncio~=3.4.3  # Library for writing concurrent code using the async/await syntax
asyncpg~=0.29.0  # Asynchronous PostgreSQL client library
aiohttp~=3.9.3  # HTTP client/server framework for asynchronous programming
python-dotenv~=0.21.0  # Library for managing environment variables in a .env file
aioconsole~=0.7.0  # Asynchronous interactive console
tenacity~=8.2.2  # Retrying library for Python
pytest~=7.4.0  # Testing framework
aioredis~=2.0.1  # Asynchronous Redis client
cachetools~=4.2.2  # Caching library
pydantic~=2.5.3  # Data validation and settings management using Python type hints
moviepy~=2.0.0.dev2  # Video editing library
urllib3~=2.0.7  # HTTP client library
jupyter~=1.0.0  # Interactive computing and data visualization
pip~=24.0  # Package installer for Python
attrs~=23.1.0  # Library for defining classes without boilerplate code
cryptography~=42.0.2  # Cryptographic library
tornado~=6.3.3  # Web framework and asynchronous networking library
Jinja2~=3.1.3  # Template engine for Python
filelock~=3.13.1  # Platform-independent file locking library
redis~=5.1.0b4  # Redis client
Django~=4.2.11  # Web framework for perfectionists with deadlines
asgiref~=3.7.2  # ASGI specification
sqlparse~=0.4.4  # SQL parsing library
psycopg2~=2.9.9  # PostgreSQL database adapter for Python
colorama~=0.4.6  # Cross-platform colored terminal text
PyYAML~=6.0.1  # YAML parser and emitter
numpy~=1.26.4  # Scientific computing library
click~=8.1.7  # Command line interface creation kit
ipython~=8.18.1  # Powerful interactive Python shell
typing_extensions~=4.9.0  # Backported and experimental type hints
gunicorn~=21.2.0  # WSGI HTTP server for UNIX
packaging~=23.1  # Core utilities for Python packages
greenlet~=3.0.1  # Light-weight microthreads for Python
whitenoise~=6.6.0  # Simplified static file serving for Django
pathlib~=1.0.1  # Object-oriented filesystem paths
panel~=1.3.8  # Interactive web dashboarding for Python
distro~=1.8.0  # Linux OS information
wheel~=0.41.2  # Built-package format for Python
pyOpenSSL~=24.1.0  # Python wrapper for OpenSSL
docutils~=0.18.1  # Documentation utilities for Python
sphinx~=5.0.2  # Documentation generator
PyYAML~=6.0.1  # YAML parser and emitter
tblib~=1.7.0  # Traceback serialization library
validators~=0.18.2  # Data validation library
brotli~=1.0.9  # Compression algorithm

# Additional dependencies
typing~=3.7.4.3  # Type hinting support
logger~=1.4  # Logging library
self~=2020.12.3  # Dependency for self
backoff~=2.2.1  # Retrying library with exponential backoff
pycryptodomex~=3.20.0  # Cryptographic library
Flask~=3.0.2  # Micro web framework
selenium~=4.18.1  # Web browser automation
